add: codeit base 코드 적용 - API 설계에 정답은 없지만, 이어지는 요구사항과 미션을 원활히 수행하기 위해 제공된 API 스펙에 맞추어 구현해주세요. - 특히, 심화 요구사항에서 제공되는 프론트엔드 코드는 제공된 API 스펙을 준수해야 연동할 수 있습니다.
라고 요구사항에 나와있지만 제가 볼때는 사실 답이 정해진 느낌입니다. 그래서 지금 이참에 모든 코드를 베이스코드로 적용하였습니다.

+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicUserStatusService.java
@@ -1,82 +1,84 @@
 package com.sprint.mission.discodeit.service.basic;
 
+import com.sprint.mission.discodeit.dto.request.UserStatusCreateRequest;
+import com.sprint.mission.discodeit.dto.request.UserStatusUpdateRequest;
 import com.sprint.mission.discodeit.entity.UserStatus;
 import com.sprint.mission.discodeit.repository.UserRepository;
 import com.sprint.mission.discodeit.repository.UserStatusRepository;
 import com.sprint.mission.discodeit.service.UserStatusService;
-import java.util.List;
-import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
-@Service
+import java.time.Instant;
+import java.util.List;
+import java.util.NoSuchElementException;
+import java.util.UUID;
+
 @RequiredArgsConstructor
+@Service
 public class BasicUserStatusService implements UserStatusService {
 
-  private final UserRepository userRepository;
   private final UserStatusRepository userStatusRepository;
+  private final UserRepository userRepository;
 
   @Override
-  public UserStatus addUserStatus(UUID userId) {
-    userRepository.findById(userId)
-        .orElseThrow(() -> new IllegalArgumentException("BasicUserStatusService: User not found"));
-    userStatusRepository.findByUserId(userId)
-        .ifPresent(userStatus -> {
-          throw new IllegalArgumentException("BasicUserStatusService: UserStatus already exists");
-        });
+  public UserStatus create(UserStatusCreateRequest request) {
+    UUID userId = request.userId();
 
-    UserStatus userStatus = new UserStatus(userId);
-    return userStatusRepository.save(userStatus)
-        .orElseThrow(() -> new RuntimeException("Failed to save UserStatus"));
+    if (!userRepository.existsById(userId)) {
+      throw new NoSuchElementException("User with id " + userId + " does not exist");
+    }
+    if (userStatusRepository.findByUserId(userId).isPresent()) {
+      throw new IllegalArgumentException("UserStatus with id " + userId + " already exists");
+    }
+
+    Instant lastActiveAt = request.lastActiveAt();
+    UserStatus userStatus = new UserStatus(userId, lastActiveAt);
+    return userStatusRepository.save(userStatus);
   }
 
   @Override
-  public UserStatus getUserStatusById(UUID id) {
-    return userStatusRepository.findById(id)
+  public UserStatus find(UUID userStatusId) {
+    return userStatusRepository.findById(userStatusId)
         .orElseThrow(
-            () -> new IllegalArgumentException("BasicUserStatusService: UserStatus not found"));
-  }
-
-
-  @Override
-  public List<UserStatus> getAllUserStatus() {
-    return userStatusRepository.findAll();
+            () -> new NoSuchElementException("UserStatus with id " + userStatusId + " not found"));
   }
 
   @Override
-  public void deleteAllUserStatus() {
-    userStatusRepository.deleteAll();
+  public List<UserStatus> findAll() {
+    return userStatusRepository.findAll().stream()
+        .toList();
   }
 
   @Override
-  public UserStatus updateUserStatus(UUID id) {
-    UserStatus userStatus = userStatusRepository.findById(id)
+  public UserStatus update(UUID userStatusId, UserStatusUpdateRequest request) {
+    Instant newLastActiveAt = request.newLastActiveAt();
+
+    UserStatus userStatus = userStatusRepository.findById(userStatusId)
         .orElseThrow(
-            () -> new IllegalArgumentException("BasicUserStatusService: UserStatus not found"));
+            () -> new NoSuchElementException("UserStatus with id " + userStatusId + " not found"));
+    userStatus.update(newLastActiveAt);
 
-    userStatus.updateLastOnlineTime();
-
-    return userStatusRepository.save(userStatus)
-        .orElseThrow(() -> new RuntimeException("Failed to update UserStatus"));
+    return userStatusRepository.save(userStatus);
   }
 
   @Override
-  public UserStatus updateUserStatusByUserId(UUID userId) {
-    userRepository.findById(userId)
-        .orElseThrow(() -> new IllegalArgumentException("BasicUserStatusService: User not found"));
+  public UserStatus updateByUserId(UUID userId, UserStatusUpdateRequest request) {
+    Instant newLastActiveAt = request.newLastActiveAt();
 
     UserStatus userStatus = userStatusRepository.findByUserId(userId)
         .orElseThrow(
-            () -> new IllegalArgumentException("BasicUserStatusService: UserStatus not found"));
+            () -> new NoSuchElementException("UserStatus with userId " + userId + " not found"));
+    userStatus.update(newLastActiveAt);
 
-    userStatus.updateLastOnlineTime();
-
-    return userStatusRepository.save(userStatus)
-        .orElseThrow(() -> new RuntimeException("Failed to update UserStatus"));
+    return userStatusRepository.save(userStatus);
   }
 
   @Override
-  public void deleteUserStatus(UUID userStatusId) {
+  public void delete(UUID userStatusId) {
+    if (!userStatusRepository.existsById(userStatusId)) {
+      throw new NoSuchElementException("UserStatus with id " + userStatusId + " not found");
+    }
     userStatusRepository.deleteById(userStatusId);
   }
 }
